+++ packages/atlas/src/components/_buttons/Button/index.ts
@@ -1,1 +1,3 @@
 export * from './Button'
+
+export type { ButtonBaseStyleProps, ButtonVariant, ButtonSize } from '../Button/Button.styles'
