--- packages/atlas/src/components/_buttons/ButtonBase/index.ts
@@ -1,3 +1,0 @@
-export * from './ButtonBase'
-
-export type { ButtonBaseStyleProps, ButtonVariant, ButtonSize } from './ButtonBase.styles'
